refactor: 렌더 함수 폴더 분리
