+++ README.md
@@ -4,16 +4,16 @@
 
 - [ ] 영화 목록의 1페이지를 불러오며 더보기 버튼을 누르면 그 다음의 영화 목록을 불러 올 수 있다.
 - [ ] 페이지 끝에 도달한 경우에는 더보기 버튼을 화면에 출력하지 않는다.
-- [ ] 영화는 한 번의 요청당 20개씩 영화 목록을 보여준다.
-- [ ] 영화 목록을 불러오는 동안 Skeleton UI 를 보여준다
-  - [ ] Skeleton UI는 템플릿으로 제공되는 파일 이외로 자유롭게 구현할 수 있다.
+- [x] 영화는 한 번의 요청당 20개씩 영화 목록을 보여준다.
+- [x] 영화 목록을 불러오는 동안 Skeleton UI 를 보여준다
+  - [x] Skeleton UI는 템플릿으로 제공되는 파일 이외로 자유롭게 구현할 수 있다.
 
 ## 검색
 
-- [ ] 영화 검색 API를 이용하여 내가 보고 싶은 영화를 검색할 수 있다.
+- [x] 영화 검색 API를 이용하여 내가 보고 싶은 영화를 검색할 수 있다.
   - [x] 엔터키를 눌러 검색할 수 있다
   - [x] 검색 버튼을 클릭하여 검색할 수 있다
-  - [ ] 영화 목록 조회와 같이 검색한 결과에 한해 정보를 보여주는 화면의 요구사항은 동일하다
+  - [x] 영화 목록 조회와 같이 검색한 결과에 한해 정보를 보여주는 화면의 요구사항은 동일하다
 
 ## 오류
 
@@ -48,7 +48,7 @@
 
 ### 메인 페이지
 
-- [ ] 최초 접근 시 인기순 영화 20개 렌더링
+- [x] 최초 접근 시 인기순 영화 20개 렌더링
 
 ### 더보기 버튼
 
@@ -56,7 +56,7 @@
 
 ### 검색
 
-- [ ] 검색 버튼 클릭 or 엔터 입력 시 검색 api 호출
+- [x] 검색 버튼 클릭 or 엔터 입력 시 검색 api 호출
 
 ### 로고
 
@@ -69,3 +69,4 @@
 - [ ] getElementById로 가져온 html 요소 타입 좁혀주는 유틸
 - [ ] TMDBError class 사용에 대해 논의
 - [ ] 돔 요소 삭제 시 eventHandler 제거
+- [ ] 빈 keyword 검색 시 기본 UI로 변경
