+++ src/main.ts
@@ -1,5 +1,7 @@
 import { getPopularMovies, Movie } from "./apis/movie/api";
 import { getSearchedMovies } from "./apis/search/api";
+import { renderResultSectionContent } from "./dom/render/renderResultSectionContent";
+import { renderThumbnailList } from "./dom/render/renderThumbnailList";
 
 const searchInput = document.getElementById(
   "search-input",
@@ -28,7 +30,7 @@ const handleSearch = async (keyword: string) => {
     movies = searchResult.results;
     isLoading = false;
     renderResultSectionContent(isLoading, isError, movies);
-    renderThumbnailList();
+    renderThumbnailList(movies);
   }
 };
 
@@ -40,74 +42,15 @@ if (searchInput && searchButton) {
   });
 }
 
-const renderResultSectionContent = (
-  isLoading: boolean,
-  isError: boolean,
-  movies: Movie[],
-) => {
-  console.log(isLoading, isError, movies);
-
-  if (isLoading) {
-    const skeletonList = document.getElementById("skeleton-list");
-    skeletonList?.classList.remove("hidden");
-    return;
-  }
-
-  if (isError) {
-    const errorContainer = document.getElementById("error-container");
-    errorContainer?.classList.remove("hidden");
-    return;
-  }
-
-  if (movies.length > 0) {
-    const thumbnailList = document.getElementById("thumbnail-list");
-    thumbnailList?.classList.remove("hidden");
-    return;
-  }
-
-  const emptyContainer = document.getElementById("empty-container");
-  emptyContainer?.classList.remove("hidden");
-};
-
 // TODO: getPopularMovies try-catch 감싸 -> 에러 핸들링
 // TODO 이미지 없는 거 대체 이미지
 renderResultSectionContent(isLoading, isError, movies);
 const popularMovies = await getPopularMovies({ language: "ko-KR" });
-console.log(popularMovies);
 movies = popularMovies.results;
 
-const renderThumbnailList = () => {
-  const thumbnailList = document.getElementById("thumbnail-list");
-  if (thumbnailList) {
-    console.log("hi");
-    const lis = movies.map(
-      (movie) => `<li id="movie-${movie.id}">
-                  <div class="item">
-                    <img
-                      class="thumbnail"
-                      src="${import.meta.env.VITE_TMDB_IMAGE_BASE_URL}/w500${movie.poster_path}"
-                      alt="${movie.title} 포스터"
-                    />
-                    <div class="item-desc">
-                      <p class="rate">
-                        <img src="./images/star_empty.png" class="star" /><span
-                          >${movie.vote_average}</span
-                        >
-                      </p>
-                      <strong>${movie.title}</strong>
-                    </div>
-                  </div>
-                </li>`,
-    );
-    thumbnailList.innerHTML = lis.join("");
-    isLoading = false;
-  }
-};
+renderThumbnailList(movies);
+isLoading = false;
 
-renderThumbnailList();
-
-const skeletonList = document.getElementById("skeleton-list");
-skeletonList?.classList.add("hidden");
 renderResultSectionContent(isLoading, isError, movies);
 
 // 진입 -> renderResultSectionContent 호출 -> api 호출
